+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/grpc/referece/GrpcMessageFieldReference.kt
@@ -1,5 +1,8 @@
 package io.kanro.idea.plugin.protobuf.grpc.referece
 
+import com.bybutter.sisyphus.protobuf.primitives.Timestamp
+import com.bybutter.sisyphus.protobuf.primitives.now
+import com.bybutter.sisyphus.protobuf.primitives.string
 import com.intellij.codeInsight.completion.InsertHandler
 import com.intellij.codeInsight.completion.InsertionContext
 import com.intellij.codeInsight.lookup.LookupElement
@@ -17,9 +20,11 @@ import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufFieldDefinition
 import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufGroupDefinition
 import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufMessageDefinition
 import io.kanro.idea.plugin.protobuf.lang.psi.filterItem
+import io.kanro.idea.plugin.protobuf.lang.psi.firstItemOrNull
 import io.kanro.idea.plugin.protobuf.lang.psi.jsonName
 import io.kanro.idea.plugin.protobuf.lang.psi.primitive.structure.ProtobufFieldLike
 import io.kanro.idea.plugin.protobuf.lang.psi.repeated
+import io.kanro.idea.plugin.protobuf.lang.support.WellknownTypes
 import kotlin.math.max
 
 class GrpcMessageFieldReference(field: JsonStringLiteral) : PsiReferenceBase<JsonStringLiteral>(field), GrpcReference {
@@ -38,14 +43,23 @@ class GrpcMessageFieldReference(field: JsonStringLiteral) : PsiReferenceBase<Jso
         val obj = property.parent as? JsonObject ?: return ArrayUtilRt.EMPTY_OBJECT_ARRAY
         val existsFields = obj.propertyList.map { it.name }.toSet()
 
+        if (type is ProtobufMessageDefinition) {
+            val message = type.qualifiedName().toString()
+            if (message == WellknownTypes.ANY) {
+                val field = type.firstItemOrNull<ProtobufFieldLike> { it.name() == "type_url" }
+                    ?: return ArrayUtilRt.EMPTY_OBJECT_ARRAY
+                return listOfNotNull(lookupForTypeUrl(field)).toTypedArray()
+            } else if (message in WellknownTypes.types) {
+                return ArrayUtilRt.EMPTY_OBJECT_ARRAY
+            }
+        }
+
         return when (type) {
             is ProtobufMessageDefinition -> {
-                type.filterItem<ProtobufFieldLike> { true }
-                    .mapNotNull { lookupFor(it, existsFields) }.toTypedArray()
+                type.filterItem<ProtobufFieldLike> { true }.mapNotNull { lookupFor(it, existsFields) }.toTypedArray()
             }
             is ProtobufGroupDefinition -> {
-                type.filterItem<ProtobufFieldLike> { true }
-                    .mapNotNull { lookupFor(it, existsFields) }.toTypedArray()
+                type.filterItem<ProtobufFieldLike> { true }.mapNotNull { lookupFor(it, existsFields) }.toTypedArray()
             }
             else -> ArrayUtilRt.EMPTY_OBJECT_ARRAY
         }
@@ -65,26 +79,57 @@ class GrpcMessageFieldReference(field: JsonStringLiteral) : PsiReferenceBase<Jso
         }?.withInsertHandler(JsonFieldSmartInsertHandler(element))
     }
 
+    private fun lookupForTypeUrl(element: ProtobufFieldLike): LookupElementBuilder? {
+        return element.lookup("@type")?.withPresentableText("@type")?.withLookupString("@type")
+            ?.appendTailText(" (type_url)", true)?.withInsertHandler(SmartInsertHandler("\": \"\"", -1, true))
+    }
+
     override fun handleElementRename(newElementName: String): PsiElement {
         return element
     }
 }
 
 private class JsonFieldSmartInsertHandler(private val element: ProtobufFieldLike) : InsertHandler<LookupElement> {
+    companion object {
+        private val OBJECT_INSERT_HANDLER = SmartInsertHandler("\": {}", -1)
+        private val LIST_INSERT_HANDLER = SmartInsertHandler("\": []", -1)
+        private val BOOLEAN_INSERT_HANDLER = SmartInsertHandler("\": ", 0, true)
+        private val STRING_INSERT_HANDLER = SmartInsertHandler("\": \"\"", -1)
+        private val STRING_REF_INSERT_HANDLER = SmartInsertHandler("\": \"\"", -1, true)
+        private val NUMBER_INSERT_HANDLER = SmartInsertHandler("\": ")
+    }
+
     override fun handleInsert(context: InsertionContext, item: LookupElement) {
         val insertHandler = when (element) {
             is ProtobufFieldDefinition -> {
                 if (element.repeated()) {
-                    SmartInsertHandler("\": []", -1)
+                    LIST_INSERT_HANDLER
                 } else {
-                    when (element.typeName.reference?.resolve()) {
-                        is ProtobufMessageDefinition -> SmartInsertHandler("\": {}", -1)
-                        is ProtobufEnumDefinition -> SmartInsertHandler("\": \"\"", -1, true)
+                    when (val type = element.typeName.reference?.resolve()) {
+                        is ProtobufMessageDefinition -> {
+                            val name = type.qualifiedName().toString()
+                            if (name in WellknownTypes.types) {
+                                when (name) {
+                                    WellknownTypes.ANY, WellknownTypes.STRUCT -> OBJECT_INSERT_HANDLER
+                                    WellknownTypes.LIST_VALUE -> LIST_INSERT_HANDLER
+                                    WellknownTypes.BOOL_VALUE -> BOOLEAN_INSERT_HANDLER
+                                    WellknownTypes.TIMESTAMP -> SmartInsertHandler(
+                                        "\": \"${
+                                            Timestamp.now().string()
+                                        }\"", -1
+                                    )
+                                    WellknownTypes.DURATION -> SmartInsertHandler("\": \"1.0s\"", -1)
+                                    WellknownTypes.FIELD_MASK, WellknownTypes.STRING_VALUE, WellknownTypes.BYTES_VALUE -> STRING_INSERT_HANDLER
+                                    else -> NUMBER_INSERT_HANDLER
+                                }
+                            } else OBJECT_INSERT_HANDLER
+                        }
+                        is ProtobufEnumDefinition -> STRING_REF_INSERT_HANDLER
                         else -> {
                             when (element.typeName.text) {
-                                "string", "bytes" -> SmartInsertHandler("\": \"\"", -1)
-                                "bool" -> SmartInsertHandler("\": ", 0, true)
-                                else -> SmartInsertHandler("\": ")
+                                "string", "bytes" -> STRING_INSERT_HANDLER
+                                "bool" -> BOOLEAN_INSERT_HANDLER
+                                else -> NUMBER_INSERT_HANDLER
                             }
                         }
                     }
@@ -92,12 +137,12 @@ private class JsonFieldSmartInsertHandler(private val element: ProtobufFieldLike
             }
             is ProtobufGroupDefinition -> {
                 if (element.repeated()) {
-                    SmartInsertHandler("\": []", -1)
+                    LIST_INSERT_HANDLER
                 } else {
-                    SmartInsertHandler("\": {}", -1)
+                    OBJECT_INSERT_HANDLER
                 }
             }
-            else -> SmartInsertHandler("\": {}", -1)
+            else -> OBJECT_INSERT_HANDLER
         }
         insertHandler.handleInsert(context, item)
     }