+++ CHANGELOG.md
@@ -6,9 +6,10 @@
 
 ### Added
 
-- Support send gRPC request via HTTP client plugin #102
-- Use the common folding setting to config import folding #100
+- Support send gRPC request via HTTP client plugin [#102](https://github.com/devkanro/intellij-protobuf-plugin/issues/102)
+- Use the common folding setting to config import folding [#100](https://github.com/devkanro/intellij-protobuf-plugin/issues/100)
 - Inject Markdown to Protobuf line comment blocks
+- Support disable decompile from go descriptor [#99](https://github.com/devkanro/intellij-protobuf-plugin/issues/99)
 
 ## [1.5.12]